extract repeatPassword component
